+++ packages/eas-cli/global-setup.ts
@@ -1,3 +1,5 @@
 module.exports = async () => {
   process.env.TZ = 'UTC';
+  process.env.EAS_NO_VCS = '1';
+  process.env.EAS_PROJECT_ROOT = '/app';
 };