+++ packages/eas-cli/src/credentials/android/actions/__tests__/RemoveFcm-test.ts
@@ -4,7 +4,7 @@ import {
   testAndroidAppCredentialsFragment,
 } from '../../../__tests__/fixtures-android';
 import { createCtxMock } from '../../../__tests__/fixtures-context';
-import { getAppLookupParamsFromContext } from '../BuildCredentialsUtils';
+import { getAppLookupParamsFromContextAsync } from '../BuildCredentialsUtils';
 import { RemoveFcm } from '../RemoveFcm';
 
 jest.mock('../../../../prompts');
@@ -31,14 +31,14 @@ describe(RemoveFcm, () => {
         ),
       },
     });
-    const appLookupParams = getAppLookupParamsFromContext(ctx);
+    const appLookupParams = await getAppLookupParamsFromContextAsync(ctx);
     const removeFcmApiKeyAction = new RemoveFcm(appLookupParams);
     await removeFcmApiKeyAction.runAsync(ctx);
     expect(ctx.android.deleteFcmAsync as any).toHaveBeenCalledTimes(1);
   });
   it('errors in Non-Interactive Mode', async () => {
     const ctx = createCtxMock({ nonInteractive: true });
-    const appLookupParams = getAppLookupParamsFromContext(ctx);
+    const appLookupParams = await getAppLookupParamsFromContextAsync(ctx);
     const removeFcmApiKeyAction = new RemoveFcm(appLookupParams);
     await expect(removeFcmApiKeyAction.runAsync(ctx)).rejects.toThrowError();
   });