+++ packages/eas-cli/src/project/android/applicationId.ts
@@ -11,7 +11,7 @@ import Log, { learnMore } from '../../log';
 import { getProjectConfigDescription, getUsername } from '../../project/projectUtils';
 import { promptAsync } from '../../prompts';
 import { ensureLoggedInAsync } from '../../user/actions';
-import { resolveWorkflow } from '../workflow';
+import { resolveWorkflowAsync } from '../workflow';
 
 const INVALID_APPLICATION_ID_MESSAGE = `Invalid format of Android applicationId. Only alphanumeric characters, '.' and '_' are allowed, and each '.' must be followed by a letter.`;
 
@@ -19,18 +19,18 @@ export async function ensureApplicationIdIsDefinedForManagedProjectAsync(
   projectDir: string,
   exp: ExpoConfig
 ): Promise<string> {
-  const workflow = resolveWorkflow(projectDir, Platform.ANDROID);
+  const workflow = await resolveWorkflowAsync(projectDir, Platform.ANDROID);
   assert(workflow === Workflow.MANAGED, 'This function should be called only for managed projects');
 
   try {
-    return getApplicationId(projectDir, exp);
+    return await getApplicationIdAsync(projectDir, exp);
   } catch (err) {
     return await configureApplicationIdAsync(projectDir, exp);
   }
 }
 
-export function getApplicationId(projectDir: string, exp: ExpoConfig): string {
-  const workflow = resolveWorkflow(projectDir, Platform.ANDROID);
+export async function getApplicationIdAsync(projectDir: string, exp: ExpoConfig): Promise<string> {
+  const workflow = await resolveWorkflowAsync(projectDir, Platform.ANDROID);
   if (workflow === Workflow.GENERIC) {
     warnIfAndroidPackageDefinedInAppConfigForGenericProject(projectDir, exp);